Refactor CodeGenTest.hpp
Common parts for testing compiler/optimizers are extracted out.
Those parts are put into a new class called CompilerUnitTest.hpp.
From now on, CompilerUnitTest.hpp can be used as a base class for compiler unit tests.
